requirements.txt
requests
spotipy
python-dotenv
deezer-python
deezer-oauth-cli